chore: 데이터 시딩, 타입 변경

+++ src/main/java/com/factory/management_service/domain/entity/LotEntity.java
@@ -37,7 +37,7 @@ public class LotEntity {
     private String lotGrade;
 
     @Column(precision = 22, scale = 5)
-    private BigDecimal productQty;
+    private Integer productQty;
 
     @Column(length = 50)
     private String productionType;
